requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic>=2.10.0
rich==13.7.0
anthropic>=0.40.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
websocket-client>=1.6.0
eventlet>=0.33.0
pyjwt>=2.8.0
bcrypt>=4.0.1
watchdog>=3.0.0
